Renamed AppwriteSpotifyProfile to SpotifyProfile
