+++ spoti-friends/src/spotify/services/ProfileServiceManager.swift
@@ -25,10 +25,10 @@ class ProfileServiceManager {
     /// Retrieves a Spotify Profile from the database using their Spotify ID.
     ///
     /// - Parameter spotifyId: The Spotify ID of the Spotify Profile to retrieve.
-    /// - Returns: The `AppwriteSpotifyProfile` object representing the profile.
+    /// - Returns: The `SpotifyProfile` object representing the profile.
     ///
     /// This method asynchronously fetches a Spotify Profile from the database by their Spotify ID.
-    func getProfileFromDB(withSpotifyId spotifyId: String) async throws -> AppwriteSpotifyProfile? {
+    func getProfileFromDB(withSpotifyId spotifyId: String) async throws -> SpotifyProfile? {
         do {
             return try await profileService.getProfileFromDB(withSpotifyId: spotifyId)
         } catch {
