+++ spoti-friends/src/spotify/spotifyAPI/helpers/GetListOfUsersFriends.swift
@@ -10,10 +10,10 @@ extension SpotifyAPI {
     ///   - data: Data object returned from `buddylist` endpoint.
     ///
     /// - Returns: A list of the user's friends as `SpotifyProfile` objects.
-    internal func convertDataToFriendList(_ data: Data) throws -> [AppwriteSpotifyProfile] {
+    internal func convertDataToFriendList(_ data: Data) throws -> [SpotifyProfile] {
         do {
             let buddylistResponseObject = try JSONDecoder().decode(BuddylistResponseObject.self, from: data)
-            var friendList: [AppwriteSpotifyProfile] = []
+            var friendList: [SpotifyProfile] = []
             for friendObject in buddylistResponseObject.friends {
                 friendList.append(friendObjectToSpotifyProfile(friendObject))
             }
@@ -30,14 +30,14 @@ extension SpotifyAPI {
     ///   - friendObject: The friend object to convert.
     ///
     /// - Returns: The friend as a `SpotifyProfile` object.
-    private func friendObjectToSpotifyProfile(_ friendObject: BuddylistFriendObject) -> AppwriteSpotifyProfile {
-        let spotifyId = AppwriteSpotifyProfile.getSpotifyIdFromUri(spotifyUri: friendObject.user.uri)
+    private func friendObjectToSpotifyProfile(_ friendObject: BuddylistFriendObject) -> SpotifyProfile {
+        let spotifyId = SpotifyProfile.getSpotifyIdFromUri(spotifyUri: friendObject.user.uri)
         let spotifyUri = friendObject.user.uri
         let displayName = friendObject.user.name
         let image = friendObject.user.imageUrl ?? ""
 //        spotifyProfile.currentOrMostRecentTrack = getCurrentOrMostRecentTrackForFriend(friendObject)
         
-        return AppwriteSpotifyProfile(spotifyId: spotifyId, spotifyUri: spotifyUri, displayName: displayName, image: image)
+        return SpotifyProfile(spotifyId: spotifyId, spotifyUri: spotifyUri, displayName: displayName, image: image)
     }
     
     /// Creates and returns the `CurrentOrMostRecentTrack` for the friend from the `/buddylist` endpoint response object.
